Simplify graphics using a palette swapping shader.
The previous method of saving PNG files via drawing tile data to
a canvas was kinda slow, and likely buggy (browsers can decided to
compress color data, which is bad if you want to use colors as indices
in a palette). Instead, I make a texture directly from the tile data,
and wrote a basic shader to do the color mapping. I was using a plugin
to do this before, but this shader does much simpler work than the color
replace one.